% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{coef.IL}
\alias{coef.IL}
\title{coef.IL return the estimated IL parameters from an IL object}
\usage{
\method{coef}{IL}(object, ...)
}
\arguments{
\item{object}{an IL object from the fitIL function}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
a vector of the estimated IL parameters from an IL object
}
\description{
coef.IL return the estimated IL parameters from an IL object
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
coef(ans)
}
