% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{dobootIL}
\alias{dobootIL}
\title{dobootIL conducts a bootstrap analysis on an IL object}
\usage{
dobootIL(x, reps = 1000)
}
\arguments{
\item{x}{an IL object such as produced by fitIL}

\item{reps}{how many boostraps to produce, 1000 should be regarded
as a minimum but that will usually take a litle time to run}
}
\value{
a class bootIL object containing the bootstrap replicate
    parameter estimates and the median plus lower and upp 95% bounds
    of each parameter.
}
\description{
dobootIL conducts a bootstrap analysis on an IL object
    so as to characterize the uncertainty in the analysis. In the
    process it produces a 'bootIL' object. This contains 'reps'
    replicate sets of parameter estimates which can be plotted
    using the S3 plot method for bootIL objects.
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
boot <- dobootIL(ans,reps=10)
str(boot)
}
