% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{fitIL.data.frame}
\alias{fitIL.data.frame}
\title{fitIL.data.frame S3 method to apply fitIL to a matrix}
\usage{
\method{fitIL}{data.frame}(x, ...)
}
\arguments{
\item{x}{a data.frame containing Lt and DL. The names must be exact}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
if x is a data.frame this fits an IL curve
}
\description{
fitIL.data.frame S3 method to apply fitIL to a matrix
}
\examples{
data(midg)
ans <- fitIL(midg,siteid=1,outliers=TRUE,sitename="midground")
str(ans)
}
