% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{fitIL.default}
\alias{fitIL.default}
\title{fitIL.default the default S3 method to apply fitIL}
\usage{
\method{fitIL}{default}(x, y, siteid = 0, outliers = FALSE, sitename = "", ...)
}
\arguments{
\item{x}{a vector containing Lt the names must be exact}

\item{y}{a vector containing DL the names must be exact}

\item{siteid}{a numeric site identifier, default = 0}

\item{outliers}{default = FALSE, should outliers be identified and
then omitted?}

\item{sitename}{an alphanumeric identifier for the site}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
returns an IL curve fit as a class IL object
}
\description{
fitIL.default the default S3 method to apply fitIL
}
\examples{
data(midg)
ans <- fitIL(midg$Lt,midg$DL,sitename="Middle Ground")
str(ans)
}
