% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{fitIL.list}
\alias{fitIL.list}
\title{fitIL.list S3 method to apply fitIL to a matrix}
\usage{
\method{fitIL}{list}(x, ...)
}
\arguments{
\item{x}{a list containing Lt and DL in the first two components}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
if x is a list this fits an IL curve
}
\description{
fitIL.list S3 method to apply fitIL to a matrix
}
\examples{
data(midg)
fitIL(list(midg$Lt,midg$DL))
}
