% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{fitIL.matrix}
\alias{fitIL.matrix}
\title{fitIL.matrix S3 method to apply fitIL to a matrix}
\usage{
\method{fitIL}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix containing Lt and DL in the first two columns}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
if x is a matrix this fits an IL curve
}
\description{
fitIL.matrix S3 method to apply fitIL to a matrix
}
\examples{
data(midg)
fitIL(cbind(midg$Lt,midg$DL))
}
