% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{fitIL}
\alias{fitIL}
\title{fitIL S3 generic method for fitting an IL curve to data}
\usage{
fitIL(x, ...)
}
\arguments{
\item{x}{a data type containing, first, the Initial lengths, Lt,
and, second, the growth increments, DL. These can be passed
as separate vectors, as a matrix, a list, or a data.frame.}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
returns a fitted IL curve of class IL
}
\description{
fitIL is an overloaded front-end function used to fit
    an IL curve to different data types containing at least two
    vectors of data required: x=Lt (the initial lengths) and y=DL
    (the corresponding growth increments). These data types include
    a matrix of x,y, a list of x, y, and the default, which is two
    vectors x=Lt, and y=DL. This is an S3 generic method, whose
    only task is to direct the analysis to the correct method by
    class of data input or go  off to the default.
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=FALSE,)
str(ans)
}
