% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{getdyn}
\alias{getdyn}
\title{getdyn estimates age-structure given tagging growth fit}
\usage{
getdyn(x, initL = 2, maxage = 24)
}
\arguments{
\item{x}{the IL object from fitIL}

\item{initL}{the length of post-larval anaimals, default = 2.0mm}

\item{maxage}{the maximum age+1 for the calculations}
}
\value{
a list containing the length at maximum age, the age at
    maturity, the length-at-age, and gradient between the
    interquartile growth increments
}
\description{
getdyn uses the tagging growth fit, the implied length
    at age 0+, and the maximum age, to estimate the length-at-age
    and the growth increment at average length at age. It enables
    an estimation of the age at maturity and the length of animals
    at age maxage-1
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
getdyn(ans,maxage=30)
}
