% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{initpars}
\alias{initpars}
\title{initpars estimates the starting parameters for fitting an IL curve}
\usage{
initpars(Ltin, DLin)
}
\arguments{
\item{Ltin}{a vector of initial lengths to which the IL is to be fitted}

\item{DLin}{a vector of growth increments to which the IL is to be
fitted}
}
\value{
a vector of four parameters for the IL (MaxDL, L50, L95, MaxSig)
}
\description{
initpars estimates the starting parameters for fitting
    an IL curve. This uses approximations and rules of thumb.
}
\examples{
  data(midg)
  initpars(midg$Lt,midg$DL)
}
