% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{midg}
\alias{midg}
\title{midg is an abalone tagging data-set from the Actaeons}
\format{
A data.frame of abalone tagging data
\describe{
  \item{RecapL}{the length at recapture}
  \item{Lt}{the length at tagging}
  \item{Dt}{The time interval between tagging and recapture, in this
      instance they are all listed as 1 year}
  \item{DL}{the growth increment in mm}
}
}
\source{
Thanks to the Institute of Marine and Antarctic Science,
    which is part of the University of Tasmania, and especially to
    Dr Craig Mundy, leader of the Abalone Group, for permission to use
    this data collected in 2003.
}
\description{
midg is a tagging data-set for blacklip abalone
    (\emph{Haliotis rubra}) from the middle ground in the Actaeons
    in Tasmania's Block 13. All individuals were recaptured in 2003,
    the site number was 478, at Latitude -43.54 longitude 146.99,
    and there are 347 observations. All Dt = 1 year.
}
\section{Subjects}{

 \itemize{
   \item growth curves
   \item inverse logistic, von Bertalanffy, Gompertz
   \item Static model fitting
 }
}

\examples{
 data(midg)
 head(midg,20)
 oldpar <- par(no.readonly=TRUE)
 plot(midg$Lt,midg$DL,type="p",pch=16,cex=1.0,xlim=c(5,180))
 abline(h=0,col=1)
 par(oldpar)
}
