% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{plot.bootIL}
\alias{plot.bootIL}
\title{plot.bootIL an S3 method to plot the bootstrap results}
\usage{
\method{plot}{bootIL}(x, col = 0, font = 7, ...)
}
\arguments{
\item{x}{the bootIL object from the dobootIL function}

\item{col}{the col of the histogram cells, default = 0 = empty}

\item{font}{default=7, bold serif, 1 is sans-serif}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
nothing but oit does generate a plot
}
\description{
plot.bootIL an S3 method for plotting the results of
    bootsrapping a inverse logistic model fit to a set of tagging\
    data. A 2 x 2 plot of histograms is generated illustrating the
    distribution of the bootstrap replicate parameter estimates.
    This has no net effect on the par defnition.
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
boot <- dobootIL(ans,reps=10)
plot(boot)
}
