% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{plot.IL}
\alias{plot.IL}
\title{plot.IL an S3 method to plot an IL object}
\usage{
\method{plot}{IL}(x, ...)
}
\arguments{
\item{x}{an IL object from the fitIL function}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
nothing but it does generatea plot
}
\description{
plot.IL is an S3 method for plotting the results of
    fitting an inverse logistic curve to a set of tagging data.
    it plot the data and the model fit, with curves depicted the
    90th percentile spread and, if outliers was set to TRUE in the
    call to fitIL, it identifies outliers as red dots, and removes
    those points from the calculation. It plots the residuals, the
    rate of change in DL, to find the turnover length, and the
    density of data by initial length
}
\examples{
data(midg)
ans <- fitIL(midg,outliers=TRUE,sitename="Middle Ground")
plot(ans)
}
