% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{print.IL}
\alias{print.IL}
\title{print.IL an S3 method ofr IL objects}
\usage{
\method{print}{IL}(x, ...)
}
\arguments{
\item{x}{an IL objects, as generated by fitIL}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
nothing but it prints a few details to the console
}
\description{
print.IL is an S3 method for IL objects, as generated
    by the fitIL function
}
\examples{
data(midg)
ans <- fitIL(midg,sitename="midgrd")
print(ans)
}
