% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{summary.IL}
\alias{summary.IL}
\title{summary.IL an S3 method for the summary generic for IL objects}
\usage{
\method{summary}{IL}(object, filename = "", ...)
}
\arguments{
\item{object}{an IL object generated by fitIL}

\item{filename}{to send output of summary to a file, default="" = to console}

\item{...}{the ellipsis is for any remaining parameters}
}
\value{
nothing but prints a summary ot the console
}
\description{
summary.IL is an S3 method for the summary generic for
    IL objects, as generated by the fitIL function.
}
\examples{
data(midg)
ans <- fitIL(midg,filename="")
summary(ans)
}
