% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funIL.R
\name{summary.IL}
\alias{summary.IL}
\title{summary.IL an S3 method for the summary generic for IL objects}
\usage{
\method{summary}{IL}(object, console = TRUE)
}
\arguments{
\item{object}{an IL object generated by fitIL}

\item{console}{print the summary to console, default=TRUE}
}
\value{
a vector of character strings. and can prints summary to the console
}
\description{
summary.IL is an S3 method for the summary generic for
    IL objects, as generated by the fitIL function.
}
\examples{
data(midg)
ans <- fitIL(midg,sitename="middleground")
out <- summary(ans,console=TRUE)
}
